+++ lib/boot_completed.dart
@@ -68,11 +68,21 @@ Future<String> _readPrivateFileAsString(String fileName) async {
   return await File((await _getPrivateDocsPath()) + fileName).readAsString();
 }
 
-//invoke android method to save entry point handle
-Future<bool> _saveDartEntryPointToAndroid() async {
-  const String _channelName = 'org.thebus.boot_completed.SaveDartEntryPoint';
-  const MethodChannel _channel = MethodChannel(_channelName, JSONMethodCodec());
+MethodChannel _pluginMethodChannel() =>
+    MethodChannel('org.thebus.boot_completed.BootCompletedPlugin.MethodChannel', JSONMethodCodec());
 
-  return await _channel.invokeMethod<bool>(
+//invoke android method to save entry point handle
+Future<bool> _saveDartEntryPointToAndroid() async =>
+  await _pluginMethodChannel().invokeMethod<bool>(
       'SaveDartEntryPoint', <dynamic>[await _getPluginBootCompletedHandle()]);
-}
+
+///If true, will disable automatic execution on boot of the function
+///set by setBootCompletedFunction
+///
+///you can then call (kotlin) BootCompletePlugin.handleBootCompleted elsewhere
+///
+///mainly useful if you have other boot receivers
+///and want to execute them in a particular order
+Future<bool> setDeferExecution(bool deferExecution) async =>
+    await _pluginMethodChannel().invokeMethod<bool>(
+        'DeferExcecution', <dynamic>[deferExecution]);